add exam ex for teacher
- add ex for teacher: Institute of Technology, https://ocw.mit.edu/courses/mechanical-engineering/
2-71-optics-spring-2009/assignments/MIT2_71S09_usol1.pdf)
Task 1: Data Types, Input/Output, Arithmetic Operations, and the math

Module A rescuer must reach a drowning person as quickly as possible to prevent his death.
Since the rescuer is on the shore and the drowning person is in the water, the rescuer must
cover part of the distance on the sand and the rest by swimming. The direction of movement
chosen by the rescuer determines the total distance he must cover, since the speed of movement
in the water is less than the speed of movement on land by a certain constant value.
You need to write a program that would calculate the time it takes the rescuer to reach the drowning person.
The notations used and the geometric construction illustrating the problem are shown in Figure 1.